Add files via upload
In this dataset, there is no explicit target column indicating whether someone has diabetes or not. Therefore, for the purpose of analysis and model prediction, I created a new target column named "Diabetes" based on the value of Fasting Blood Glucose.
According to the guidelines from the American Diabetes Association (ADA) and the World Health Organization (WHO), a fasting blood glucose level of ≥126 mg/dL is used as the threshold for diagnosing diabetes. This information was obtained from several official sources, including [this link](https://www.mayoclinic.org/diseases-conditions/diabetes/diagnosis-treatment/drc-20371451?utm_source=chatgpt.com).

Using this threshold, I classify patients as diabetic (1) if their fasting blood glucose level is ≥126, and as non-diabetic (0) if it is lower. This target column was very useful for training predictive models such as Logistic Regression and others.